MEDIA-02: a peer-local stream stop/refusal must not retire the video writer
FrameStream erased every StreamWriteError/StreamOpeningError into a
static string, and every failure retired the per-(client, source)
writer. That over-classifies: on wtransport 0.7.1, StreamWriteError::
Stopped(code) and StreamOpeningError::Refused mean the peer abandoned
ONE stream, not the connection — so a single browser/relay decision
to drop a late frame silenced that source until reconnect while the
connection and other sources stayed healthy.

Open/write failures now carry a typed StreamError classification:
PeerLocal { phase, code } for Stopped/Refused/Closed (one-frame loss:
count, log with client/source/phase/peer-code, continue) and
ConnectionFatal for NotConnected/QuicProto/open-request errors (retire
the writer). The absolute per-frame deadline and explicit RESET_STREAM
from #129 are unchanged; a failed frame is never retried. Logs and
counters now distinguish deadline stalls (total_stalls, stream_reset)
from peer-local drops (total_peer_drops, phase, peer_code) from
connection-fatal loss.

Virtual-time tests: a peer-stopped write and a refused open each lose
one frame while the next frame opens and finishes and the writer
survives; a connection-fatal write and a connection-fatal open each
retire the writer. Tests moved to a frame_writer/tests submodule
(500-line limit).

Closes #132
