+++ hosts/session-host/src/runtime/media/frame_writer.rs
@@ -9,6 +9,7 @@ use pilotage_session::ClientKey;
 use tokio::sync::mpsc;
 use tokio::time::Instant;
 use tracing::{debug, error, warn};
+use wtransport::error::{StreamOpeningError, StreamWriteError};
 use wtransport::{Connection, SendStream, VarInt};
 
 use super::{EncodedFrame, now_ns};
@@ -29,6 +30,38 @@ const FRAME_WRITE_DEADLINE: Duration = Duration::from_secs(2);
 /// partial frame regardless of the code.
 const STALL_RESET_CODE: u32 = 1;
 
+/// Why an open, write, or finish failed, classified so a peer's decision
+/// to abandon ONE stream does not retire the whole source writer.
+enum StreamError {
+    /// The peer stopped or refused this stream alone (`Stopped`,
+    /// `Refused`), or it was already closed locally. The connection and
+    /// every other source are unaffected — this is one-frame loss.
+    PeerLocal {
+        /// The phase that surfaced it (`open`, `write`, `finish`).
+        phase: &'static str,
+        /// The peer's application error code, when it carried one.
+        code: Option<u64>,
+    },
+    /// Connection-level loss or a protocol failure (`NotConnected`,
+    /// `QuicProto`, or an open-request `ConnectionError`): the writer must
+    /// retire — no further frame can be delivered on this connection.
+    ConnectionFatal(&'static str),
+}
+
+/// Classifies a wtransport write/finish error: only `NotConnected` and
+/// `QuicProto` are connection-fatal; `Stopped`/`Closed` are peer-local.
+fn classify_write(error: &StreamWriteError, phase: &'static str) -> StreamError {
+    match error {
+        StreamWriteError::Stopped(code) => StreamError::PeerLocal {
+            phase,
+            code: Some(code.into_inner()),
+        },
+        StreamWriteError::Closed => StreamError::PeerLocal { phase, code: None },
+        StreamWriteError::NotConnected => StreamError::ConnectionFatal("not connected"),
+        StreamWriteError::QuicProto => StreamError::ConnectionFatal("QUIC protocol error"),
+    }
+}
+
 /// One per-frame outbound stream. `write_all`/`finish` are the clean send
 /// path; `reset` is the explicit RESET_STREAM a deadline-exceeded frame
 /// needs. Dropping a wtransport/Quinn `SendStream` attempts a graceful
@@ -38,9 +71,9 @@ const STALL_RESET_CODE: u32 = 1;
 /// because the writer runs as a spawned task on a multi-threaded runtime.
 trait FrameStream {
     /// Writes the whole buffer, awaiting flow-control credit.
-    fn write_all(&mut self, buf: &[u8]) -> impl Future<Output = Result<(), &'static str>> + Send;
+    fn write_all(&mut self, buf: &[u8]) -> impl Future<Output = Result<(), StreamError>> + Send;
     /// Finishes the stream cleanly (graceful FIN).
-    fn finish(&mut self) -> impl Future<Output = Result<(), &'static str>> + Send;
+    fn finish(&mut self) -> impl Future<Output = Result<(), StreamError>> + Send;
     /// Resets the stream (RESET_STREAM); a no-op on an already-closed one.
     fn reset(&mut self);
 }
@@ -50,20 +83,20 @@ trait FrameChannel {
     /// The stream this channel opens.
     type Stream: FrameStream;
     /// Opens one host-initiated uni stream.
-    fn open(&self) -> impl Future<Output = Result<Self::Stream, &'static str>> + Send;
+    fn open(&self) -> impl Future<Output = Result<Self::Stream, StreamError>> + Send;
 }
 
 impl FrameStream for SendStream {
-    async fn write_all(&mut self, buf: &[u8]) -> Result<(), &'static str> {
+    async fn write_all(&mut self, buf: &[u8]) -> Result<(), StreamError> {
         SendStream::write_all(self, buf)
             .await
-            .map_err(|_| "payload write failed")
+            .map_err(|e| classify_write(&e, "write"))
     }
 
-    async fn finish(&mut self) -> Result<(), &'static str> {
+    async fn finish(&mut self) -> Result<(), StreamError> {
         SendStream::finish(self)
             .await
-            .map_err(|_| "stream finish failed")
+            .map_err(|e| classify_write(&e, "finish"))
     }
 
     fn reset(&mut self) {
@@ -76,12 +109,20 @@ impl FrameStream for SendStream {
 impl FrameChannel for Connection {
     type Stream = SendStream;
 
-    async fn open(&self) -> Result<SendStream, &'static str> {
-        self.open_uni()
+    async fn open(&self) -> Result<SendStream, StreamError> {
+        // The open-request error is connection-level; the opening error can
+        // be a peer refusal of this stream alone.
+        let opening = self
+            .open_uni()
             .await
-            .map_err(|_| "uni stream request failed")?
-            .await
-            .map_err(|_| "uni stream failed to open")
+            .map_err(|_| StreamError::ConnectionFatal("uni stream request failed"))?;
+        opening.await.map_err(|e| match e {
+            StreamOpeningError::Refused => StreamError::PeerLocal {
+                phase: "open",
+                code: None,
+            },
+            StreamOpeningError::NotConnected => StreamError::ConnectionFatal("not connected"),
+        })
     }
 }
 
@@ -108,17 +149,23 @@ enum FrameOutcome {
     /// had opened and was explicitly reset, false when the open itself
     /// timed out (there was no stream to reset).
     Stalled { reset: bool },
-    /// The open or write failed outright; the connection is going away.
-    Failed(&'static str),
+    /// The peer stopped or refused this stream alone; the connection and
+    /// other sources are healthy — one-frame loss, keep writing.
+    PeerLocal {
+        /// The phase (`open`, `write`, `finish`) that surfaced it.
+        phase: &'static str,
+        /// The peer's application error code, when it carried one.
+        code: Option<u64>,
+    },
+    /// Connection-level loss or protocol failure; retire the writer.
+    ConnectionFatal(&'static str),
 }
 
 /// Drains the handoff channel, delivering one frame per stream under a
 /// single absolute per-frame deadline that covers BOTH opening the stream
-/// and writing it. A frame that exceeds its deadline costs one frame and
-/// the writer proceeds — a mid-write timeout resets the opened stream
-/// (RESET_STREAM), an open timeout simply skips (no stream exists). An
-/// outright open/write failure means the connection is going away and ends
-/// the writer.
+/// and writing it. A frame lost to a deadline or a peer-local stop/refusal
+/// costs one frame and the writer proceeds; only connection-level loss
+/// retires the writer.
 async fn drain_frames<C: FrameChannel>(
     client: ClientKey,
     source_id: u8,
@@ -127,6 +174,7 @@ async fn drain_frames<C: FrameChannel>(
     start: Instant,
 ) {
     let mut stalls: u64 = 0;
+    let mut peer_drops: u64 = 0;
     while let Some(frame) = frames.recv().await {
         // Stamp publication at the moment of write, distinct from the receive
         // stamp taken at dequeue, so a consumer can separate host queueing
@@ -158,10 +206,21 @@ async fn drain_frames<C: FrameChannel>(
                     "video frame exceeded its deadline; continuing with the next frame"
                 );
             }
-            FrameOutcome::Failed(reason) => {
-                // A failed open/write means the connection is going away; stop
-                // writing video to it. The connection task's own teardown
-                // deregisters this client.
+            FrameOutcome::PeerLocal { phase, code } => {
+                peer_drops = peer_drops.wrapping_add(1);
+                warn!(
+                    client = client.as_u64(),
+                    source_id,
+                    phase,
+                    peer_code = code,
+                    total_peer_drops = peer_drops,
+                    "peer stopped or refused this video stream; the connection is healthy, \
+                     continuing with the next frame"
+                );
+            }
+            FrameOutcome::ConnectionFatal(reason) => {
+                // Connection-level loss; stop writing video to it. The
+                // connection task's own teardown deregisters this client.
                 debug!(
                     client = client.as_u64(),
                     source_id, reason, "video writer stopping"
@@ -176,7 +235,9 @@ async fn drain_frames<C: FrameChannel>(
 /// absolute `deadline`. An open that exceeds the deadline yields a stall
 /// with no stream to reset; a write that exceeds it explicitly resets the
 /// opened stream (RESET_STREAM), retained across the timed region so the
-/// reset can fire rather than a dropped-future FIN.
+/// reset can fire rather than a dropped-future FIN. Typed open/write
+/// failures pass their peer-local vs connection-fatal classification
+/// through unchanged.
 async fn deliver_frame<C: FrameChannel>(
     channel: &C,
     deadline: Instant,
@@ -185,15 +246,26 @@ async fn deliver_frame<C: FrameChannel>(
 ) -> FrameOutcome {
     let mut stream = match tokio::time::timeout_at(deadline, channel.open()).await {
         Ok(Ok(stream)) => stream,
-        Ok(Err(reason)) => return FrameOutcome::Failed(reason),
+        Ok(Err(error)) => return error.into_outcome(),
         Err(_elapsed) => return FrameOutcome::Stalled { reset: false },
     };
     match tokio::time::timeout_at(deadline, write_body(&mut stream, tag, body)).await {
         Ok(Ok(())) => FrameOutcome::Sent,
-        Ok(Err(reason)) => FrameOutcome::Failed(reason),
+        Ok(Err(error)) => error.into_outcome(),
         Err(_elapsed) => {
             stream.reset();
             FrameOutcome::Stalled { reset: true }
+        }
+    }
+}
+
+impl StreamError {
+    /// Maps a stream error to its frame outcome, preserving the peer-local
+    /// vs connection-fatal classification.
+    fn into_outcome(self) -> FrameOutcome {
+        match self {
+            StreamError::PeerLocal { phase, code } => FrameOutcome::PeerLocal { phase, code },
+            StreamError::ConnectionFatal(reason) => FrameOutcome::ConnectionFatal(reason),
         }
     }
 }
@@ -203,232 +275,11 @@ async fn write_body<S: FrameStream>(
     stream: &mut S,
     tag: u8,
     body: &[u8],
-) -> Result<(), &'static str> {
+) -> Result<(), StreamError> {
     stream.write_all(&[tag]).await?;
     stream.write_all(body).await?;
     stream.finish().await
 }
 
 #[cfg(test)]
-#[allow(clippy::expect_used, clippy::panic)]
-mod tests {
-    use std::sync::Arc;
-    use std::sync::atomic::{AtomicU32, Ordering};
-
-    use pilotage_adapter_api::{
-        CalibrationId, CameraId, CaptureClockMapping, MeasurementClock, MeasurementStamp,
-        SourceIncarnation, SourceIntegrity, SourceRole, VideoCaptureStamp,
-    };
-    use pilotage_session::ClientKey;
-    use tokio::sync::mpsc;
-    use tokio::time::Instant;
-
-    use super::{EncodedFrame, FrameChannel, FrameStream, drain_frames};
-
-    fn capture_stamp() -> VideoCaptureStamp {
-        VideoCaptureStamp {
-            stamp: MeasurementStamp {
-                role: SourceRole::VideoCapture,
-                integrity: SourceIntegrity::Unprotected,
-                source_id: 1,
-                source_incarnation: SourceIncarnation::new([5; 16]),
-                source_epoch: 0,
-                sequence: 3,
-                acquired_at_ns: 999,
-                clock: MeasurementClock::Simulation,
-            },
-            camera_id: CameraId(1),
-            calibration_id: CalibrationId::NONE,
-            mapping: CaptureClockMapping::identity(MeasurementClock::Simulation),
-        }
-    }
-
-    fn encoded_frame() -> EncodedFrame {
-        EncodedFrame {
-            jpeg: Arc::new(vec![0xFF, 0xD8, 0xFF, 0xD9]),
-            capture: capture_stamp(),
-            received_at_ns: 0,
-        }
-    }
-
-    /// Shared call tallies a test asserts against.
-    #[derive(Default)]
-    struct Tally {
-        opened: AtomicU32,
-        finished: AtomicU32,
-        reset: AtomicU32,
-    }
-
-    /// How a mock stream behaves once opened.
-    #[derive(Clone, Copy)]
-    enum Write {
-        /// `write_all` never completes (a wedged consumer).
-        Stall,
-        /// `write_all` returns an error (connection going away).
-        Fail,
-        /// Writes and finishes normally.
-        Ok,
-    }
-
-    /// How a mock `open()` behaves.
-    #[derive(Clone, Copy)]
-    enum Open {
-        /// `open()` never completes (the peer's stream allowance is full).
-        Stall,
-        /// `open()` yields a stream with the given write behavior.
-        Ready(Write),
-    }
-
-    struct MockStream {
-        write: Write,
-        tally: Arc<Tally>,
-    }
-
-    impl FrameStream for MockStream {
-        async fn write_all(&mut self, _buf: &[u8]) -> Result<(), &'static str> {
-            match self.write {
-                Write::Stall => {
-                    std::future::pending::<()>().await;
-                    Ok(())
-                }
-                Write::Fail => Err("payload write failed"),
-                Write::Ok => Ok(()),
-            }
-        }
-
-        async fn finish(&mut self) -> Result<(), &'static str> {
-            self.tally.finished.fetch_add(1, Ordering::SeqCst);
-            Ok(())
-        }
-
-        fn reset(&mut self) {
-            self.tally.reset.fetch_add(1, Ordering::SeqCst);
-        }
-    }
-
-    /// Hands out opens from a per-open script; opens beyond the script are
-    /// `Ready(Write::Ok)`.
-    struct MockChannel {
-        script: Vec<Open>,
-        tally: Arc<Tally>,
-    }
-
-    impl FrameChannel for MockChannel {
-        type Stream = MockStream;
-
-        async fn open(&self) -> Result<MockStream, &'static str> {
-            let n = self.tally.opened.fetch_add(1, Ordering::SeqCst) as usize;
-            match self
-                .script
-                .get(n)
-                .copied()
-                .unwrap_or(Open::Ready(Write::Ok))
-            {
-                Open::Stall => {
-                    std::future::pending::<()>().await;
-                    unreachable!("a stalled open never resolves")
-                }
-                Open::Ready(write) => Ok(MockStream {
-                    write,
-                    tally: self.tally.clone(),
-                }),
-            }
-        }
-    }
-
-    async fn queue(frames: usize) -> mpsc::Receiver<EncodedFrame> {
-        let (tx, rx) = mpsc::channel(frames.max(1));
-        for _ in 0..frames {
-            tx.send(encoded_frame()).await.expect("frame queues");
-        }
-        drop(tx);
-        rx
-    }
-
-    /// A stalled write must reset its stream exactly once (RESET_STREAM,
-    /// not a dropped FIN) and the next frame must open its own stream and
-    /// finish cleanly. Virtual time fires the deadline without real waiting.
-    #[tokio::test(start_paused = true)]
-    async fn a_stalled_write_resets_once_and_the_next_frame_proceeds() {
-        let mut rx = queue(2).await;
-        let tally = Arc::new(Tally::default());
-        let channel = MockChannel {
-            script: vec![Open::Ready(Write::Stall), Open::Ready(Write::Ok)],
-            tally: tally.clone(),
-        };
-
-        drain_frames(ClientKey::new(1), 0, &channel, &mut rx, Instant::now()).await;
-
-        assert_eq!(
-            tally.reset.load(Ordering::SeqCst),
-            1,
-            "the stalled frame's stream is reset exactly once"
-        );
-        assert_eq!(
-            tally.opened.load(Ordering::SeqCst),
-            2,
-            "the next frame opens its own stream"
-        );
-        assert_eq!(
-            tally.finished.load(Ordering::SeqCst),
-            1,
-            "the second frame finishes cleanly and is not reset"
-        );
-    }
-
-    /// A stalled OPEN also costs one frame: the per-frame deadline covers
-    /// opening, so an open that never completes is skipped (nothing to
-    /// reset) and the next frame opens its own stream and finishes.
-    #[tokio::test(start_paused = true)]
-    async fn a_stalled_open_is_skipped_and_the_next_frame_proceeds() {
-        let mut rx = queue(2).await;
-        let tally = Arc::new(Tally::default());
-        let channel = MockChannel {
-            script: vec![Open::Stall, Open::Ready(Write::Ok)],
-            tally: tally.clone(),
-        };
-
-        drain_frames(ClientKey::new(1), 0, &channel, &mut rx, Instant::now()).await;
-
-        assert_eq!(
-            tally.opened.load(Ordering::SeqCst),
-            2,
-            "the next frame opens its own stream after the stalled open"
-        );
-        assert_eq!(
-            tally.finished.load(Ordering::SeqCst),
-            1,
-            "the second frame finishes cleanly"
-        );
-        assert_eq!(
-            tally.reset.load(Ordering::SeqCst),
-            0,
-            "a stalled open has no stream to reset"
-        );
-    }
-
-    /// An outright open/write error ends the writer: the connection is going
-    /// away, so the frame after the failed one is never opened.
-    #[tokio::test(start_paused = true)]
-    async fn a_write_error_ends_the_writer() {
-        let mut rx = queue(2).await;
-        let tally = Arc::new(Tally::default());
-        let channel = MockChannel {
-            script: vec![Open::Ready(Write::Fail), Open::Ready(Write::Ok)],
-            tally: tally.clone(),
-        };
-
-        drain_frames(ClientKey::new(1), 0, &channel, &mut rx, Instant::now()).await;
-
-        assert_eq!(
-            tally.opened.load(Ordering::SeqCst),
-            1,
-            "the writer stops after the failed frame; the next is never opened"
-        );
-        assert_eq!(
-            tally.reset.load(Ordering::SeqCst),
-            0,
-            "an outright failure is not a reset"
-        );
-    }
-}
+mod tests;
